movierecommendationsystem/requirements.txt
streamlit==1.51.0
requests==2.32.5
python-dotenv==1.2.1
pandas>=2.0.0
scikit-learn>=1.0.0
numpy>=1.20.0
altair>=4.0.0,<6.0.0,!=5.4.0,!=5.4.1
blinker>=1.4.0
cachetools>=4.0.0
click>=7.0.0
gitpython>=3.0.7
protobuf>=3.20.0
pydeck>=0.8.0b4
tenacity>=8.1.0
toml>=0.10.1
typing-extensions>=4.4.0
watchdog>=2.1.5
mysql-connector-python>=8.0.33
bcrypt>=4.0.0




